+++ src/Fields/Base.php
@@ -4,13 +4,28 @@ namespace MVRK\Icenberg\Fields;
 
 class Base
 {
-
+    /**
+     * Wrap the field in html
+     *
+     * @param mixed $content
+     * @param mixed $classes
+     * @param mixed $layout
+     * @param mixed $tag
+     * @return string
+     */
     public function wrap($content, $classes, $layout, $tag)
     {
         return "<{$tag} class='block--{$this->unSnake($layout)}__{$this->unSnake($classes)}'>{$content}</{$tag}>";
     }
 
-
+    /**
+     * Wrap content in a html list
+     *
+     * @param mixed $content
+     * @param mixed $classes
+     * @param mixed $layout
+     * @return string
+     */
     public function listWrap($content, $classes, $layout)
     {
         $list = "";
@@ -24,7 +39,7 @@ class Base
 
     public function unSnake($text)
     {
-        return  str_replace('_', '-', $text);
+        return str_replace('_', '-', $text);
     }
 
     public function unSpace($string)
@@ -32,18 +47,73 @@ class Base
         return str_replace([' ', '-'], '_', $string);
     }
 
-    public static function icefield($field_name)
+    /**
+     * Prepare the field when retrieved from the field object
+     *
+     * @param mixed $field_name
+     * @param mixed $id
+     * @return mixed
+     */
+    public static function icefield($field_name, $id = null)
     {
-        if (!get_sub_field($field_name) && !get_field($field_name)) {
+        $options = self::isOption($id);
+
+        if (!get_sub_field($field_name) && !get_field($field_name, $options)) {
             return;
         }
 
         if (get_sub_field($field_name)) {
             $field = get_sub_field($field_name);
-        } else if (get_field($field_name)) {
-            $field = get_field($field_name);
+        } else if (get_field($field_name, $options)) {
+            $field = get_field($field_name, $options);
         }
 
         return $field;
+    }
+
+    /**
+     * All options field objects have an ID of 0
+     *
+     * @param mixed $id
+     * @return boolean
+     */
+    public static function isOption($id)
+    {
+        if ($id === 0) {
+            return 'options';
+        } else {
+            return false;
+        }
+    }
+
+    protected static function wrapInTag($content, $item_name, $class, $tag = 'span')
+    {
+        return "<{$tag} class='{$class}__{$item_name}'>{$content}</{$tag}>";
+    }
+
+    public function preview($wp_post, $name, $layout, $tag)
+    {
+        if (!$wp_post) {
+            return false;
+        }
+
+        $permalink = esc_url(get_the_permalink($wp_post));
+        $title = get_the_title($wp_post);
+        $excerpt = get_the_excerpt($wp_post);
+        $thumbnail = get_the_post_thumbnail($wp_post);
+        $date = get_the_date('j M Y', $wp_post);
+        $class = "preview--{$this->unSnake($layout)}__{$this->unSnake($name)}";
+        $linked_title = "<a href='{$permalink}'>{$title}</a>";
+        $link = "<a href='{$permalink}'>Read More</a>";
+
+        $content = "<{$tag} class='{$class}'>";
+        $content .= self::wrapInTag($date, 'date', $class, 'div');
+        $content .= self::wrapInTag($thumbnail, 'thumbnail', $class, "div");
+        $content .= self::wrapInTag($linked_title, 'title', $class, 'h4');
+        $content .= self::wrapInTag($excerpt, 'excerpt', $class, 'div');
+        $content .= self::wrapInTag($link, 'link', $class, "div");
+        $content .= "</{$tag}>";
+
+        return $content;
     }
 }