+++ src/Fields/Forms.php
@@ -7,6 +7,10 @@ class Forms extends Base
 
     protected $titles;
 
+    protected $title;
+
+    protected $description;
+
     protected $form;
 
     public $gravity_form;
@@ -20,11 +24,8 @@ class Forms extends Base
     public function getElement($field_object, $layout, $tag)
     {
         $this->form = $field_object['value'];
-
         $this->name = $field_object['_name'];
-
         $this->layout = $layout;
-
         $this->tag = $tag;
 
         $this->getTitles();