+++ src/Fields/PageLink.php
@@ -4,13 +4,23 @@ namespace MVRK\Icenberg\Fields;
 
 class PageLink extends Base
 {
-    public function getElement($field, $layout, $tag)
+    public function getElement($field_object, $layout, $tag)
     {
-        $name = $field['_name'];
+        $name = $field_object['_name'];
 
-        $link = self::icefield($name);
+        $id = $field_object['ID'];
 
-        $wrapped = "<a class='block--{$this->unSnake($layout)}__{$this->unSnake($name)}' href='{$link}'>{$link}</a>";
+        $links = self::icefield($name, $id);
+
+        if (is_string($links)) {
+            $links = [$links];
+        }
+
+        $wrapped = '';
+
+        foreach ($links as $link) {
+            $wrapped .= "<a class='block--{$this->unSnake($layout)}__{$this->unSnake($name)}' href='{$link}'>{$link}</a>";
+        }
 
         return $wrapped;
     }