+++ src/Fields/Range.php
@@ -5,11 +5,13 @@ namespace MVRK\Icenberg\Fields;
 class Range extends Base
 {
 
-    public function getElement($field, $layout, $tag)
+    public function getElement($field_object, $layout, $tag)
     {
-        $name = $field['_name'];
+        $name = $field_object['_name'];
 
-        $content = self::icefield($name);
+        $id = $field_object['ID'];
+
+        $content = self::icefield($name, $id);
 
         $wrapped = $this->wrap($content, $name, $layout, $tag);