+++ src/Fields/Wysiwyg.php
@@ -7,15 +7,17 @@ class Wysiwyg extends Base
     /**
      * Scaffold a WYSIWYG field
      *
-     * @param object $field
+     * @param object $field_object
      * @param string $layout the current row
      * @return string
      */
-    public function getElement($field, $layout, $tag)
+    public function getElement($field_object, $layout, $tag)
     {
-        $name = $field['_name'];
+        $name = $field_object['_name'];
 
-        $content = self::icefield($name);
+        $id = $field_object['ID'];
+
+        $content = self::icefield($name, $id);
 
         $wrapped = $this->wrap($content, $name, $layout, $tag);