+++ tests/Fixtures/Fields/Repeater.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace MVRK\Icenberg\Fields;
+
+
+class Repeater
+{
+    public function getElement($field_object, $instance, $tag = 'div')
+    {
+        return "<{$tag}>Repeater Content</{$tag}>";
+    }
+}